# [policy] trigger post_action phase