--- a/gateway/conf.d/apicast.conf
+++ b/gateway/conf.d/apicast.conf
@@ -63,14 +63,8 @@ location / {
   set $service_id null;
   set $proxy_pass null;
   set $secret_token null;
-  set $resp_body null;
-  set $resp_headers null;
-
-  set $client_id null;
-  set $redirect_url null;
 
   set $backend_host 'backend';
-  set $post_action_backend_endpoint '';
   set $backend_authentication_type null;
   set $backend_authentication_value null;
   set $version '';