--- a/gateway/conf.d/apicast.conf
+++ b/gateway/conf.d/apicast.conf
@@ -48,7 +48,10 @@ location @out_of_band_authrep_action {
 
   set_by_lua $original_request_time 'return ngx.var.request_time';
 
-  content_by_lua_block { require('apicast.executor'):post_action() }
+  content_by_lua_block {
+    require('resty.ctx').apply()
+    require('apicast.executor'):post_action()
+ }
 
   log_by_lua_block {
     ngx.var.post_action_impact = ngx.var.request_time - ngx.var.original_request_time
@@ -70,12 +73,17 @@ location / {
   set $version '';
   set $real_url null;
 
+  set $ctx_ref -1;
+
   set $post_action_impact null;
   set $original_request_id null;
 
   proxy_ignore_client_abort on;
 
-  rewrite_by_lua_block { require('apicast.executor'):rewrite() }
+  rewrite_by_lua_block {
+    require('resty.ctx').stash()
+    require('apicast.executor'):rewrite()
+  }
   access_by_lua_block { require('apicast.executor'):access() }
   body_filter_by_lua_block { require('apicast.executor'):body_filter() }
   header_filter_by_lua_block { require('apicast.executor'):header_filter() }