--- a/gateway/src/apicast/policy/apicast/policy.lua
+++ b/gateway/src/apicast/policy/apicast/policy.lua
@@ -61,6 +61,10 @@ function _M:post_action()
   local request_id = ngx.var.original_request_id
   local post_action_proxy = self.post_action_proxy
 
+  if not request_id then
+    return nil, 'not initialized'
+  end
+
   if not post_action_proxy then
     return nil, 'not initialized'
   end
@@ -72,7 +76,7 @@ function _M:post_action()
   if p then
     return p:post_action()
   else
-    ngx.log(ngx.INFO, 'could not find proxy for request id: ', request_id)
+    ngx.log(ngx.ERR, 'could not find proxy for request id: ', request_id)
     return nil, 'no proxy for request'
   end
 end
@@ -94,8 +98,10 @@ function _M:access(context)
     ok, err = access()
     post_action_proxy[ngx.var.original_request_id] = p
   elseif handler then
-    ok, err = handler()
     -- no proxy because that would trigger post action
+    ngx.var.original_request_id = nil
+
+    ok, err = handler()
   end
 
   return ok, err