--- a/t/apicast-policy-chains.t
+++ b/t/apicast-policy-chains.t
@@ -65,3 +65,4 @@ running phase: access
 running phase: balancer
 running phase: header_filter
 running phase: body_filter
+running phase: post_action