By sharing ngx.ctx with the subrequest we are able to 
share the policy chain and execute the post_action phase.

It **should** be safe, but we should focus on trying this during QE process.
  

# TODO
- [x] verify performance: https://gist.github.com/mikz/23bd8896dd0f0c03b88161d089261573
- [ ] merge #547
- [ ] merge #543
- [ ] merge #544


# Performance

## Before

```
wrk --connections 10 --threads 10 --duration 60 'http://localhost:8080/?user_key=foo'
Running 1m test @ http://localhost:8080/?user_key=foo
  10 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     5.25ms    0.93ms  17.38ms   69.83%
    Req/Sec   191.24     16.50   393.00     73.23%
  114377 requests in 1.00m, 41.55MB read
Requests/sec:   1903.26
Transfer/sec:    708.06KB
```

## After

```
wrk --connections 10 --threads 10 --duration 60 'http://localhost:8080/?user_key=foo'
Running 1m test @ http://localhost:8080/?user_key=foo
  10 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     5.41ms    1.07ms  21.00ms   72.87%
    Req/Sec   185.63     18.08   333.00     83.23%
  111020 requests in 1.00m, 40.33MB read
Requests/sec:   1847.49
Transfer/sec:    687.31KB
```

There is small difference, but we need to focus on performance later on.